refactor: split bus.go into focused files
Split bus.go (1,094 lines) into 4 focused files:
- bus_options.go: busOptions struct, BusOption type, all With* functions
- bus_registry.go: global bus registry, name resolution, package-level generics
- bus_health.go: health/status types and health-checking logic
- errors.go: bus error sentinels moved here alongside handler errors
- bus.go: remaining core (Bus struct, lifecycle, messaging, registration)

Pure file reorganization with no behavior or API changes.
