+++ bus.go
@@ -6,12 +6,9 @@ import (
 	"fmt"
 	"log/slog"
 	"reflect"
-	"strings"
 	"sync"
 	"sync/atomic"
-	"time"
 
-	"github.com/rbaliyan/event/v3/health"
 	"github.com/rbaliyan/event/v3/transport"
 	"github.com/rbaliyan/event/v3/transport/message"
 	"go.opentelemetry.io/otel"
@@ -30,445 +27,11 @@ const (
 	busStopped = 0
 )
 
-// FullNameSeparator is the separator between bus name and event name in full event names.
-// Full name format: "<bus_name>://<event_name>"
-const FullNameSeparator = "://"
-
-// Global bus registry
-var busRegistry sync.Map // map[string]*Bus
-
-// GetBus returns a registered bus by name.
-// Returns nil if no bus with that name exists.
-//
-// Note: The returned bus may be closed. Use GetBusOrError() if you need
-// to ensure the bus is running, or check bus.Running() after retrieval.
-func GetBus(name string) *Bus {
-	if v, ok := busRegistry.Load(name); ok {
-		return v.(*Bus)
-	}
-	return nil
-}
-
-// GetBusOrError returns a registered bus by name, or an error if the bus
-// doesn't exist or was closed at the time of the check. This is safer than
-// GetBus() when you need to verify the bus exists and is running.
-//
-// Note: The bus could be closed between this check and subsequent use (TOCTOU).
-// All Bus methods (Send, Recv, etc.) return ErrBusClosed if called after close,
-// so callers are safe from crashes but should handle ErrBusClosed on use.
-//
-// Example:
-//
-//	bus, err := event.GetBusOrError("my-bus")
-//	if err != nil {
-//	    return err // Bus doesn't exist or is closed
-//	}
-//	// Bus was running at check time; handle ErrBusClosed on subsequent calls
-func GetBusOrError(name string) (*Bus, error) {
-	v, ok := busRegistry.Load(name)
-	if !ok {
-		return nil, fmt.Errorf("%w: %q", ErrBusNotFound, name)
-	}
-	bus := v.(*Bus)
-	if !bus.Running() {
-		return nil, fmt.Errorf("%w: %q", ErrBusClosed, name)
-	}
-	return bus, nil
-}
-
-// ListBuses returns the names of all registered buses.
-func ListBuses() []string {
-	var names []string
-	busRegistry.Range(func(key, value any) bool {
-		names = append(names, key.(string))
-		return true
-	})
-	return names
-}
-
-// parseFullName splits a full event name into bus name and event name.
-// Format: "<bus_name>://<event_name>"
-// Returns error if format is invalid.
-func parseFullName(fullName string) (busName, eventName string, err error) {
-	idx := strings.Index(fullName, FullNameSeparator)
-	if idx == -1 {
-		return "", "", fmt.Errorf("%w: missing separator %q in %q", ErrInvalidFullName, FullNameSeparator, fullName)
-	}
-	busName = fullName[:idx]
-	eventName = fullName[idx+len(FullNameSeparator):]
-	if busName == "" {
-		return "", "", fmt.Errorf("%w: empty bus name in %q", ErrInvalidFullName, fullName)
-	}
-	if eventName == "" {
-		return "", "", fmt.Errorf("%w: empty event name in %q", ErrInvalidFullName, fullName)
-	}
-	return busName, eventName, nil
-}
-
-// Get retrieves a typed event by its full name.
-// Full name format: "<bus_name>://<event_name>"
-//
-// The type parameter T must match the type used when the event was registered.
-// Returns ErrTypeMismatch if the types don't match.
-//
-// Example:
-//
-//	event, err := event.Get[Order]("mybus://order.created")
-//	if err != nil {
-//	    log.Fatal(err)
-//	}
-//	event.Publish(ctx, Order{ID: "123"})
-func Get[T any](fullName string) (Event[T], error) {
-	busName, eventName, err := parseFullName(fullName)
-	if err != nil {
-		return nil, err
-	}
-
-	bus := GetBus(busName)
-	if bus == nil {
-		return nil, fmt.Errorf("%w: %q", ErrBusNotFound, busName)
-	}
-
-	// Get the type for T to validate
-	var zero T
-	eventType := reflect.TypeOf(zero)
-
-	ev, err := bus.getTyped(eventName, eventType)
-	if err != nil {
-		return nil, err
-	}
-	if ev == nil {
-		return nil, fmt.Errorf("%w: %q", ErrEventNotFound, eventName)
-	}
-
-	typed, ok := ev.(Event[T])
-	if !ok {
-		return nil, fmt.Errorf("%w: cannot cast event %q to requested type", ErrTypeMismatch, eventName)
-	}
-
-	return typed, nil
-}
-
-// Publish sends data to an event by its full name.
-// Full name format: "<bus_name>://<event_name>"
-//
-// The type parameter T must match the type used when the event was registered.
-//
-// Example:
-//
-//	err := event.Publish(ctx, "mybus://order.created", Order{ID: "123"})
-func Publish[T any](ctx context.Context, fullName string, data T) error {
-	ev, err := Get[T](fullName)
-	if err != nil {
-		return err
-	}
-	return ev.Publish(ctx, data)
-}
-
-// Subscribe registers a handler for an event by its full name.
-// Full name format: "<bus_name>://<event_name>"
-//
-// The type parameter T must match the type used when the event was registered.
-//
-// Example:
-//
-//	err := event.Subscribe(ctx, "mybus://order.created", func(ctx context.Context, e event.Event[Order], order Order) error {
-//	    fmt.Println("Received order:", order.ID)
-//	    return nil
-//	})
-func Subscribe[T any](ctx context.Context, fullName string, handler Handler[T], opts ...SubscribeOption[T]) error {
-	ev, err := Get[T](fullName)
-	if err != nil {
-		return err
-	}
-	return ev.Subscribe(ctx, handler, opts...)
-}
-
 // DefaultBusName is the default name for bus and consumer groups
 // Using the same name across distributed systems enables:
 // - WorkerPool mode: load balancing across instances (one receives)
 // - Broadcast mode: all instances receive messages
 var DefaultBusName = "event-bus"
-
-// Bus errors
-var (
-	ErrBusClosed         = errors.New("bus is closed")
-	ErrBusExists         = errors.New("bus already exists with this name")
-	ErrBusNotFound       = errors.New("bus not found")
-	ErrEventExists       = errors.New("event already exists")
-	ErrEventNotFound     = errors.New("event not found")
-	ErrTypeMismatch      = errors.New("event type mismatch")
-	ErrAlreadyBound      = errors.New("event already bound to another bus")
-	ErrTransportRequired = errors.New("transport is required: use WithBusTransport(channel.New()) or similar")
-	ErrInvalidFullName   = errors.New("invalid full name format, expected: <bus_name>://<event_name>")
-	ErrSchemaLoadFailed  = errors.New("failed to load schema from provider")
-)
-
-// StatusCode represents the health state of the bus
-type StatusCode string
-
-const (
-	// StatusHealthy indicates the bus is functioning normally
-	StatusHealthy StatusCode = "healthy"
-	// StatusDegraded indicates the bus is functioning but with issues
-	StatusDegraded StatusCode = "degraded"
-	// StatusUnhealthy indicates the bus is not functioning
-	StatusUnhealthy StatusCode = "unhealthy"
-)
-
-// Status contains detailed status information for the bus
-type Status struct {
-	Code       StatusCode         `json:"status"`
-	Message    string             `json:"message,omitempty"`
-	Latency    time.Duration      `json:"latency,omitempty"`
-	Details    map[string]any     `json:"details,omitempty"`
-	Components map[string]*Status `json:"components,omitempty"`
-	CheckedAt  time.Time          `json:"checked_at"`
-}
-
-// IsHealthy returns true if the status code is healthy
-func (s *Status) IsHealthy() bool {
-	return s.Code == StatusHealthy
-}
-
-// ConsumerLag is an alias for transport.ConsumerLag containing
-// information about consumer lag for an event.
-type ConsumerLag = transport.ConsumerLag
-
-// busOptions holds configuration for bus (unexported)
-type busOptions struct {
-	transport       transport.Transport
-	logger          *slog.Logger
-	tracingEnabled  bool
-	recoveryEnabled bool
-	metricsEnabled  bool
-	// Subscriber middleware stores (applied automatically to all subscribers)
-	idempotencyStore IdempotencyStore
-	poisonDetector   PoisonDetector
-	monitorStore     MonitorStore
-	// Schema provider for dynamic event configuration
-	schemaProvider SchemaProvider
-	strictSchema   bool // If true, fail registration when schema provider errors occur
-	// Outbox store for transactional event publishing
-	outboxStore OutboxStore
-}
-
-// BusOption option function for bus configuration
-type BusOption func(*busOptions)
-
-// WithTransport sets a custom transport for the bus
-func WithTransport(t transport.Transport) BusOption {
-	return func(o *busOptions) {
-		if t != nil {
-			o.transport = t
-		}
-	}
-}
-
-// WithTracing enables/disables tracing for all events on this bus
-func WithTracing(enabled bool) BusOption {
-	return func(o *busOptions) {
-		o.tracingEnabled = enabled
-	}
-}
-
-// WithRecovery enables/disables panic recovery for all events on this bus
-func WithRecovery(enabled bool) BusOption {
-	return func(o *busOptions) {
-		o.recoveryEnabled = enabled
-	}
-}
-
-// WithMetrics enables/disables metrics for all events on this bus
-func WithMetrics(enabled bool) BusOption {
-	return func(o *busOptions) {
-		o.metricsEnabled = enabled
-	}
-}
-
-// WithLogger sets a custom logger for the bus
-func WithLogger(l *slog.Logger) BusOption {
-	return func(o *busOptions) {
-		if l != nil {
-			o.logger = l
-		}
-	}
-}
-
-// WithIdempotency configures automatic idempotency checking for all subscribers.
-// When set, all event handlers will automatically skip duplicate messages.
-// This eliminates the need to manually check idempotency in each handler.
-//
-// Example:
-//
-//	store := idempotency.NewRedisStore(redisClient, time.Hour)
-//	bus, _ := event.NewBus("my-app",
-//	    event.WithBusTransport(transport),
-//	    event.WithIdempotency(store),
-//	)
-//
-//	// Subscriber is simple - no manual idempotency check needed
-//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
-//	    return processOrder(ctx, order) // Just business logic!
-//	})
-func WithIdempotency(store IdempotencyStore) BusOption {
-	return func(o *busOptions) {
-		if store != nil {
-			o.idempotencyStore = store
-		}
-	}
-}
-
-// WithPoisonDetection configures automatic poison message detection for all subscribers.
-// When set, all event handlers will automatically skip quarantined messages and track failures.
-// Messages that fail repeatedly will be quarantined and skipped until released.
-//
-// Example:
-//
-//	detector := poison.NewDetector(poison.NewRedisStore(redisClient),
-//	    poison.WithThreshold(5),
-//	    poison.WithQuarantineTime(time.Hour),
-//	)
-//	bus, _ := event.NewBus("my-app",
-//	    event.WithBusTransport(transport),
-//	    event.WithPoisonDetection(detector),
-//	)
-//
-//	// Subscriber is simple - no manual poison detection needed
-//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
-//	    return processOrder(ctx, order) // Just business logic!
-//	})
-func WithPoisonDetection(detector PoisonDetector) BusOption {
-	return func(o *busOptions) {
-		if detector != nil {
-			o.poisonDetector = detector
-		}
-	}
-}
-
-// WithMonitor configures automatic event processing monitoring for all subscribers.
-// When set, all event handlers will automatically record processing metrics including
-// start time, duration, status, and any errors.
-//
-// Example:
-//
-//	store := monitor.NewPostgresStore(db)
-//	bus, _ := event.NewBus("my-app",
-//	    event.WithTransport(transport),
-//	    event.WithMonitor(store),
-//	)
-//
-//	// Subscriber is simple - monitoring happens automatically
-//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
-//	    return processOrder(ctx, order) // Just business logic!
-//	})
-func WithMonitor(store MonitorStore) BusOption {
-	return func(o *busOptions) {
-		if store != nil {
-			o.monitorStore = store
-		}
-	}
-}
-
-// WithSchemaProvider configures a schema provider for dynamic event configuration.
-// When set, events will automatically load their configuration from the schema registry
-// when registered, ensuring all subscribers have consistent settings.
-//
-// The schema provider also enables real-time configuration updates via the Watch mechanism.
-//
-// Example:
-//
-//	// Using in-memory provider for testing
-//	provider := schema.NewMemoryProvider()
-//	bus, _ := event.NewBus("my-app",
-//	    event.WithTransport(transport),
-//	    event.WithSchemaProvider(provider),
-//	)
-//
-//	// Using PostgreSQL provider with notification callback
-//	provider := schema.NewPostgresProvider(db, func(ctx context.Context, change schema.SchemaChangeEvent) error {
-//	    return bus.publishSchemaChange(ctx, change)
-//	})
-func WithSchemaProvider(provider SchemaProvider) BusOption {
-	return func(o *busOptions) {
-		if provider != nil {
-			o.schemaProvider = provider
-		}
-	}
-}
-
-// WithStrictSchema configures strict schema loading behavior.
-// When enabled, event registration will fail if the schema provider
-// returns an error (e.g., database connection failure).
-//
-// By default (strict=false):
-//   - Schema not found: continue with event defaults (expected for new events)
-//   - Schema provider error: log warning and continue with defaults
-//
-// With strict=true:
-//   - Schema not found: continue with event defaults
-//   - Schema provider error: fail registration with ErrSchemaLoadFailed
-//
-// Enable this when schema-defined settings (timeouts, retries, feature flags)
-// are critical for correct operation and should not be silently ignored.
-//
-// Example:
-//
-//	bus, _ := event.NewBus("order-service",
-//	    event.WithSchemaProvider(provider),
-//	    event.WithStrictSchema(true), // Fail if schema provider errors
-//	)
-func WithStrictSchema(strict bool) BusOption {
-	return func(o *busOptions) {
-		o.strictSchema = strict
-	}
-}
-
-// WithOutbox configures an outbox store for transactional event publishing.
-// When set, calls to Publish() will automatically route to the outbox when
-// inside a transaction (detected via WithOutboxTx context).
-//
-// Normal publishes (outside transactions) still go directly to the transport.
-// This enables atomic "business operation + event publish" within database transactions.
-//
-// Example:
-//
-//	store := outbox.NewMongoStore(mongoClient, "events", "outbox")
-//	bus, _ := event.NewBus("my-app",
-//	    event.WithTransport(transport),
-//	    event.WithOutbox(store),
-//	)
-//
-//	// Normal publish - goes directly to transport
-//	orderEvent.Publish(ctx, order)
-//
-//	// Inside transaction - goes to outbox
-//	sess.WithTransaction(ctx, func(sessCtx mongo.SessionContext) (any, error) {
-//	    ctx := event.WithOutboxTx(sessCtx, sessCtx)
-//	    ordersCol.UpdateOne(ctx, filter, update)
-//	    return nil, orderEvent.Publish(ctx, order) // Routed to outbox!
-//	})
-func WithOutbox(store OutboxStore) BusOption {
-	return func(o *busOptions) {
-		if store != nil {
-			o.outboxStore = store
-		}
-	}
-}
-
-// newBusOptions creates options with defaults and applies provided options
-func newBusOptions(opts ...BusOption) *busOptions {
-	o := &busOptions{
-		logger:          slog.Default(),
-		tracingEnabled:  true,
-		recoveryEnabled: true,
-		metricsEnabled:  true,
-	}
-	for _, opt := range opts {
-		opt(o)
-	}
-	return o
-}
 
 // Bus is an event bus that manages events and their lifecycle
 type Bus struct {
@@ -721,166 +284,6 @@ func (b *Bus) getTyped(name string, eventType reflect.Type) (any, error) {
 		}
 		return existing, nil
 	}
-	return nil, nil
-}
-
-// Status returns detailed status information about the bus and its transport.
-// Use this to inspect the bus state for monitoring dashboards.
-// If the transport or any configured stores implement HealthChecker, their status is included.
-func (b *Bus) Status(ctx context.Context) *Status {
-	result := &Status{
-		CheckedAt:  time.Now(),
-		Details:    make(map[string]any),
-		Components: make(map[string]*Status),
-	}
-
-	// Check bus status
-	if !b.Running() {
-		result.Code = StatusUnhealthy
-		result.Message = "bus is closed"
-		result.Details["bus_name"] = b.name
-		return result
-	}
-
-	// Count events
-	b.eventMutex.RLock()
-	eventCount := len(b.events)
-	b.eventMutex.RUnlock()
-
-	result.Details["bus_name"] = b.name
-	result.Details["events"] = eventCount
-
-	// Start with healthy status
-	result.Code = StatusHealthy
-	result.Message = "bus is healthy"
-
-	// Check transport health if it implements HealthChecker
-	if hc, ok := b.transport.(transport.HealthChecker); ok {
-		transportHealth := hc.Health(ctx)
-		result.Components["transport"] = convertTransportStatus(transportHealth)
-		aggregateComponentStatus(result, "transport", transportHealth.Status)
-	}
-
-	// Check health of configured stores that implement health.Checker
-	b.checkComponentHealth(ctx, result, "idempotency_store", b.idempotencyStore)
-	b.checkComponentHealth(ctx, result, "poison_detector", b.poisonDetector)
-	b.checkComponentHealth(ctx, result, "monitor_store", b.monitorStore)
-	b.checkComponentHealth(ctx, result, "schema_provider", b.schemaProvider)
-	b.checkComponentHealth(ctx, result, "outbox_store", b.outboxStore)
-
-	return result
-}
-
-// checkComponentHealth checks if a component implements health.Checker and aggregates its status.
-func (b *Bus) checkComponentHealth(ctx context.Context, result *Status, name string, component any) {
-	if component == nil {
-		return
-	}
-	if hc, ok := component.(health.Checker); ok {
-		componentHealth := hc.Health(ctx)
-		result.Components[name] = convertHealthResult(componentHealth)
-		aggregateComponentStatus(result, name, health.Status(componentHealth.Status))
-	}
-}
-
-// aggregateComponentStatus updates the result status based on component status.
-// Uses the worst status: unhealthy > degraded > healthy
-func aggregateComponentStatus(result *Status, name string, status any) {
-	var statusCode StatusCode
-	switch s := status.(type) {
-	case transport.HealthStatus:
-		switch s {
-		case transport.HealthStatusUnhealthy:
-			statusCode = StatusUnhealthy
-		case transport.HealthStatusDegraded:
-			statusCode = StatusDegraded
-		default:
-			return // healthy, no change needed
-		}
-	case health.Status:
-		switch s {
-		case health.StatusUnhealthy:
-			statusCode = StatusUnhealthy
-		case health.StatusDegraded:
-			statusCode = StatusDegraded
-		default:
-			return // healthy, no change needed
-		}
-	default:
-		return
-	}
-
-	// Only update if new status is worse
-	if statusCode == StatusUnhealthy {
-		result.Code = StatusUnhealthy
-		result.Message = name + " is unhealthy"
-	} else if statusCode == StatusDegraded && result.Code != StatusUnhealthy {
-		result.Code = StatusDegraded
-		result.Message = name + " is degraded"
-	}
-}
-
-// convertHealthResult converts health.Result to bus Status
-func convertHealthResult(hr *health.Result) *Status {
-	if hr == nil {
-		return nil
-	}
-	return &Status{
-		Code:      StatusCode(hr.Status),
-		Message:   hr.Message,
-		Latency:   hr.Latency,
-		Details:   hr.Details,
-		CheckedAt: hr.CheckedAt,
-	}
-}
-
-// Health performs a health check suitable for health probes.
-// Returns nil if the bus is healthy, or an error describing the issue.
-func (b *Bus) Health(ctx context.Context) error {
-	status := b.Status(ctx)
-	if status.Code == StatusUnhealthy {
-		return errors.New(status.Message)
-	}
-	return nil
-}
-
-// convertTransportStatus converts transport.HealthCheckResult to bus Status
-func convertTransportStatus(th *transport.HealthCheckResult) *Status {
-	if th == nil {
-		return nil
-	}
-
-	result := &Status{
-		Code:      StatusCode(th.Status),
-		Message:   th.Message,
-		Latency:   th.Latency,
-		Details:   th.Details,
-		CheckedAt: th.CheckedAt,
-	}
-
-	// Convert nested components
-	if len(th.Components) > 0 {
-		result.Components = make(map[string]*Status, len(th.Components))
-		for k, v := range th.Components {
-			result.Components[k] = convertTransportStatus(v)
-		}
-	}
-
-	return result
-}
-
-// ConsumerLag returns consumer lag metrics for all events if the transport supports it.
-// Returns nil if the transport doesn't implement LagMonitor.
-func (b *Bus) ConsumerLag(ctx context.Context) ([]ConsumerLag, error) {
-	if !b.Running() {
-		return nil, ErrBusClosed
-	}
-
-	if lm, ok := b.transport.(transport.LagMonitor); ok {
-		return lm.ConsumerLag(ctx)
-	}
-
-	// Transport doesn't support lag monitoring
 	return nil, nil
 }
