+++ bus_options.go
@@ -1,0 +1,242 @@
+package event
+
+import (
+	"log/slog"
+
+	"github.com/rbaliyan/event/v3/transport"
+)
+
+// busOptions holds configuration for bus (unexported)
+type busOptions struct {
+	transport       transport.Transport
+	logger          *slog.Logger
+	tracingEnabled  bool
+	recoveryEnabled bool
+	metricsEnabled  bool
+	// Subscriber middleware stores (applied automatically to all subscribers)
+	idempotencyStore IdempotencyStore
+	poisonDetector   PoisonDetector
+	monitorStore     MonitorStore
+	// Schema provider for dynamic event configuration
+	schemaProvider SchemaProvider
+	strictSchema   bool // If true, fail registration when schema provider errors occur
+	// Outbox store for transactional event publishing
+	outboxStore OutboxStore
+}
+
+// BusOption option function for bus configuration
+type BusOption func(*busOptions)
+
+// WithTransport sets a custom transport for the bus
+func WithTransport(t transport.Transport) BusOption {
+	return func(o *busOptions) {
+		if t != nil {
+			o.transport = t
+		}
+	}
+}
+
+// WithTracing enables/disables tracing for all events on this bus
+func WithTracing(enabled bool) BusOption {
+	return func(o *busOptions) {
+		o.tracingEnabled = enabled
+	}
+}
+
+// WithRecovery enables/disables panic recovery for all events on this bus
+func WithRecovery(enabled bool) BusOption {
+	return func(o *busOptions) {
+		o.recoveryEnabled = enabled
+	}
+}
+
+// WithMetrics enables/disables metrics for all events on this bus
+func WithMetrics(enabled bool) BusOption {
+	return func(o *busOptions) {
+		o.metricsEnabled = enabled
+	}
+}
+
+// WithLogger sets a custom logger for the bus
+func WithLogger(l *slog.Logger) BusOption {
+	return func(o *busOptions) {
+		if l != nil {
+			o.logger = l
+		}
+	}
+}
+
+// WithIdempotency configures automatic idempotency checking for all subscribers.
+// When set, all event handlers will automatically skip duplicate messages.
+// This eliminates the need to manually check idempotency in each handler.
+//
+// Example:
+//
+//	store := idempotency.NewRedisStore(redisClient, time.Hour)
+//	bus, _ := event.NewBus("my-app",
+//	    event.WithBusTransport(transport),
+//	    event.WithIdempotency(store),
+//	)
+//
+//	// Subscriber is simple - no manual idempotency check needed
+//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
+//	    return processOrder(ctx, order) // Just business logic!
+//	})
+func WithIdempotency(store IdempotencyStore) BusOption {
+	return func(o *busOptions) {
+		if store != nil {
+			o.idempotencyStore = store
+		}
+	}
+}
+
+// WithPoisonDetection configures automatic poison message detection for all subscribers.
+// When set, all event handlers will automatically skip quarantined messages and track failures.
+// Messages that fail repeatedly will be quarantined and skipped until released.
+//
+// Example:
+//
+//	detector := poison.NewDetector(poison.NewRedisStore(redisClient),
+//	    poison.WithThreshold(5),
+//	    poison.WithQuarantineTime(time.Hour),
+//	)
+//	bus, _ := event.NewBus("my-app",
+//	    event.WithBusTransport(transport),
+//	    event.WithPoisonDetection(detector),
+//	)
+//
+//	// Subscriber is simple - no manual poison detection needed
+//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
+//	    return processOrder(ctx, order) // Just business logic!
+//	})
+func WithPoisonDetection(detector PoisonDetector) BusOption {
+	return func(o *busOptions) {
+		if detector != nil {
+			o.poisonDetector = detector
+		}
+	}
+}
+
+// WithMonitor configures automatic event processing monitoring for all subscribers.
+// When set, all event handlers will automatically record processing metrics including
+// start time, duration, status, and any errors.
+//
+// Example:
+//
+//	store := monitor.NewPostgresStore(db)
+//	bus, _ := event.NewBus("my-app",
+//	    event.WithTransport(transport),
+//	    event.WithMonitor(store),
+//	)
+//
+//	// Subscriber is simple - monitoring happens automatically
+//	orderEvent.Subscribe(ctx, func(ctx context.Context, e event.Event[Order], order Order) error {
+//	    return processOrder(ctx, order) // Just business logic!
+//	})
+func WithMonitor(store MonitorStore) BusOption {
+	return func(o *busOptions) {
+		if store != nil {
+			o.monitorStore = store
+		}
+	}
+}
+
+// WithSchemaProvider configures a schema provider for dynamic event configuration.
+// When set, events will automatically load their configuration from the schema registry
+// when registered, ensuring all subscribers have consistent settings.
+//
+// The schema provider also enables real-time configuration updates via the Watch mechanism.
+//
+// Example:
+//
+//	// Using in-memory provider for testing
+//	provider := schema.NewMemoryProvider()
+//	bus, _ := event.NewBus("my-app",
+//	    event.WithTransport(transport),
+//	    event.WithSchemaProvider(provider),
+//	)
+//
+//	// Using PostgreSQL provider with notification callback
+//	provider := schema.NewPostgresProvider(db, func(ctx context.Context, change schema.SchemaChangeEvent) error {
+//	    return bus.publishSchemaChange(ctx, change)
+//	})
+func WithSchemaProvider(provider SchemaProvider) BusOption {
+	return func(o *busOptions) {
+		if provider != nil {
+			o.schemaProvider = provider
+		}
+	}
+}
+
+// WithStrictSchema configures strict schema loading behavior.
+// When enabled, event registration will fail if the schema provider
+// returns an error (e.g., database connection failure).
+//
+// By default (strict=false):
+//   - Schema not found: continue with event defaults (expected for new events)
+//   - Schema provider error: log warning and continue with defaults
+//
+// With strict=true:
+//   - Schema not found: continue with event defaults
+//   - Schema provider error: fail registration with ErrSchemaLoadFailed
+//
+// Enable this when schema-defined settings (timeouts, retries, feature flags)
+// are critical for correct operation and should not be silently ignored.
+//
+// Example:
+//
+//	bus, _ := event.NewBus("order-service",
+//	    event.WithSchemaProvider(provider),
+//	    event.WithStrictSchema(true), // Fail if schema provider errors
+//	)
+func WithStrictSchema(strict bool) BusOption {
+	return func(o *busOptions) {
+		o.strictSchema = strict
+	}
+}
+
+// WithOutbox configures an outbox store for transactional event publishing.
+// When set, calls to Publish() will automatically route to the outbox when
+// inside a transaction (detected via WithOutboxTx context).
+//
+// Normal publishes (outside transactions) still go directly to the transport.
+// This enables atomic "business operation + event publish" within database transactions.
+//
+// Example:
+//
+//	store := outbox.NewMongoStore(mongoClient, "events", "outbox")
+//	bus, _ := event.NewBus("my-app",
+//	    event.WithTransport(transport),
+//	    event.WithOutbox(store),
+//	)
+//
+//	// Normal publish - goes directly to transport
+//	orderEvent.Publish(ctx, order)
+//
+//	// Inside transaction - goes to outbox
+//	sess.WithTransaction(ctx, func(sessCtx mongo.SessionContext) (any, error) {
+//	    ctx := event.WithOutboxTx(sessCtx, sessCtx)
+//	    ordersCol.UpdateOne(ctx, filter, update)
+//	    return nil, orderEvent.Publish(ctx, order) // Routed to outbox!
+//	})
+func WithOutbox(store OutboxStore) BusOption {
+	return func(o *busOptions) {
+		if store != nil {
+			o.outboxStore = store
+		}
+	}
+}
+
+// newBusOptions creates options with defaults and applies provided options
+func newBusOptions(opts ...BusOption) *busOptions {
+	o := &busOptions{
+		logger:          slog.Default(),
+		tracingEnabled:  true,
+		recoveryEnabled: true,
+		metricsEnabled:  true,
+	}
+	for _, opt := range opts {
+		opt(o)
+	}
+	return o
+}
